code/requirements.txt
fastapi==0.103.2
mangum==0.17.0
uvicorn==0.23.2
scikit-learn==1.1.3
python-dotenv==1.0.1
